Extract protocol hander into module
